templates層の追加と軽微なリファクタ
